icon_uploader/requirements.txt
Flask
pillow
pymysql
passlib
